Basic restructure
Begin exploring some of the ideas encountered in the functional
programming world of SICP through the (for me) more familiar world of
python and lilypond. This commit starts to looks at some of the features
of the itertools module to create sequences. Otherwise we've slightly
reorganized the previously existing scheme code.